Substantial changes to the WebDAVTreeView.  This now is the primary underlying model of the data in the system.  The WebDAVFileView is now a secondary datamodel which sends information back to the TreeView on events such as selection.
The processing has been eliminated from getChildeCount in WebDAVTreeView.
This has allowed for the capturing of the exand and selection of
the nodes via event listeners.  A further change in the node processing
occurs with changes in loadChildren, loadRemote, and finishLoadRemote.
loadRemote now acts in a thread aware way.  Previously it immediately
tried to read the input buffer after the request thread was launched.
Now, when the reponse interpretor gets the info back, the finish method is
called which loads the children onto the node.

A number of Event Listeners have been unhooked.  The purpose of these
event listeners were to keep four classes in sych: Main, WebDAVTreeView,
WebDAVFileView, and WebDAVRequestGenerator.  The goal was to eliminate the
synching problems created and to make data more manageable by putting it
in the underlying model.  Previously the lock info/token was managed
by the View and not the Model (MVC paradigm).  By using means such as
the new methods in WebDAVFileView (getSelected, getNode), information can
be set in the WebDAVRequestGenerator( via setResource, setNode) which
allows for the command operations to get the pertinant info for a
"transaction" of HTTP DAV requests to be sent.  Now commnads can get the info
from the underlying data Model, and ship it to the Request object along
with specific comman info which is supplied in Main.

An new "back-side" Event has been added.  The CopyResponseEvent and the
associated CopyResponseListener.  The purpose for this event is to allow for
a second prop find to be sent in order to reload a node in
which an actions has changed the node's structure on the server.  This
ensures some synching of the dataModel and the Server.  The processing
done in reposne to the Event is to call a propfind on the affected node,
reloading the node and updating the display.

The Events of Request and Response have the Node affected attached.  Thios
allows for the Node affected to be know on the response from the Server,
thus the node will be properly updated.

Also, the SelectionEvent now carries the real Node and real TreePath
for selected items.  This is in contrast to carrying the string
representations for these two objects.