Changed the way the Create Collection command works.  Now when a collection is selected in the right side file view, the collection is made in the selected collection.
Double click on a collection (dir) in the local file system would
not be sent properly, so now an adjustment has been made in the
Tree Slection Listener so that local diretories are properly
handled.

The Write File dialog box now saves the last directory entered.  This
should assist anyone adding multiple files from a specific directory.

The Save As dialog now checks for platfom specific info.  This is because
the overwrite alert would apear twice on Windows systems.

The Rename function now works on both collections and resources.

There were a lot of thread unsafe calls being made.  Many of the functions
now set the variables properly.

The Extra string is now explicitly set before GenerateMkCol.  This is because
not setting it would sometimes cause the wrong post processing in Main's
ResponseListener to be triggered.  This was the root cause of a bug
in which Delete would not work after a certain sequence of calls to the
server.